// src/api/generated/models/exportFormatResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { ExportFormatResponseDtoCapabilities } from './exportFormatResponseDtoCapabilities';
import type { ExportFormatResponseDtoFormatClass } from './exportFormatResponseDtoFormatClass';

export interface ExportFormatResponseDto {
  id: string;
  label: string;
  fileExtension: string;
  mediaType: string;
  formatClass: ExportFormatResponseDtoFormatClass;
  capabilities: ExportFormatResponseDtoCapabilities;
}